chore(organization): enforce governed script topology
