+++ archive/scripts/legacy-organization/README.md
@@ -1,0 +1,11 @@
+# Legacy Organization Scripts
+
+This directory retains retired organization helpers and stale duplicate script
+implementations as historical evidence. These files are not active operator
+entrypoints.
+
+Use the governed validators instead:
+
+- `./.auto-organize.sh --check --verbose`
+- `python3 scripts/governance/check_script_topology.py --verbose`
+- `python3 scripts/governance/check_docs_structure.py --all`
