+++ archive/scripts/legacy-organization/organize_docs_root_legacy.sh
