+++ docs/historical/script-audits/README.md
@@ -1,0 +1,6 @@
+# Script Audit History
+
+This directory stores point-in-time script reports and review summaries that no
+longer belong in the active `scripts/` tree. Keep current operator guidance in
+`scripts/README.md`, `scripts/setup/README.md`, and
+`docs/governance/REPO_ORGANIZATION.md`.
