+++ scripts/README.md
@@ -19,6 +19,7 @@ The repository uses an automated file organization system to maintain a clean st
 - **`pre_commit_hook.sh`** - Compatibility wrapper for running the unified quality gate directly
 - **`utilities/pre-commit-quality-check.py`** - Compatibility quality gate used by `pre_commit_hook.sh`
 - **`setup/pre-commit-check.sh`** - Standalone root-file-placement check used by the quality gate
+- **`governance/check_script_topology.py`** - Governed script-placement and wrapper-contract validator
 - **`utilities/verify_organization.sh`** - Verify repository organization
 
 ### Quick Start
@@ -42,10 +43,10 @@ Most scripts can be run directly:
 
 ```bash
 # Shell scripts
-./scripts/setup/install_models.sh
+./scripts/setup/install_da3_runtime.sh
 
 # Python scripts
-python scripts/utilities/verify_setup.py
+python scripts/setup/install_models.py --dry-run
 ```
 
 ## Contributing
@@ -56,6 +57,7 @@ When adding new scripts:
 2. Make executable: `chmod +x your_script.sh`
 3. Add header comments explaining purpose and usage
 4. Update this README if adding a new category
-5. Test with `./.auto-organize.sh --dry-run` to verify organization
+5. Keep top-level `scripts/*.py` paths as compatibility wrappers only
+6. Test with `./.auto-organize.sh --dry-run` to verify organization
 
 See [REPO_ORGANIZATION.md](../docs/governance/REPO_ORGANIZATION.md) for complete guidelines.
